+++ build.gradle
@@ -19,4 +19,8 @@ dependencies {
     tasks.withType(JavaCompile) {    options.encoding = 'UTF-8'}
     // https://mvnrepository.com/artifact/jstl/jstl
     implementation group: 'jstl', name: 'jstl', version: '1.2'
+    implementation group: 'mysql', name: 'mysql-connector-java', version: '8.0.23'
+    compileOnly group: 'javax.servlet', name: 'javax.servlet-api', version: '4.0.1'
+    implementation group: 'javax.servlet', name: 'jstl', version: '1.2'
 }
+
