+++ src/main/java/controller/AccountServlet.java
@@ -1,0 +1,4 @@
+package controller;
+
+public class AccountServlet {
+}
