+++ src/main/java/controller/CatalogiServlet.java
@@ -1,0 +1,4 @@
+package controller;
+
+public class CatalogiServlet {
+}
