+++ src/main/java/repository/INewSpaperRepository.java
@@ -1,0 +1,18 @@
+package repository;
+
+import model.NewSpaper;
+
+import java.sql.SQLException;
+import java.util.List;
+
+public interface INewSpaperRepository {
+//    void add (NewSpaper newSpaper);
+//
+//    NewSpaper finByID(int id);
+//
+//    void update (NewSpaper newSpaper);
+//
+//    void delete(NewSpaper newSpaper);
+s
+    List<NewSpaper> listNew();
+}
