--- src/main/java/repository/impl/BaseRepository.java
@@ -1,24 +1,0 @@
-package repository.impl;
-
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
-
-public class BaseRepository {
-    private static String jdbcURL = "jdbc:mysql://localhost:3306/newspaper";
-    private static String jdbcUsername = "root";
-    private static String jdbcPassword = "0964297433DOng";
-    private static Connection connection;
-
-    public static Connection getConnection() {
-        if (connection == null) {
-            try {
-                Class.forName("com.mysql.jdbc.Driver");
-                connection = DriverManager.getConnection(jdbcURL, jdbcUsername, jdbcPassword);
-            } catch (SQLException | ClassNotFoundException e) {
-                e.printStackTrace();
-            }
-        }
-        return connection;
-    }
-}
