+++ src/main/java/repository/newSpaperimpl/BaseRepository.java
@@ -1,0 +1,24 @@
+package repository.newSpaperimpl;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+public class BaseRepository {
+    private static String jdbcURL = "jdbc:mysql://localhost:3306/newspaper";
+    private static String jdbcUsername = "root";
+    private static String jdbcPassword = "0964297433Dong";
+    private static Connection connection;
+
+    public static Connection getConnection() {
+        if (connection == null) {
+            try {
+                Class.forName("com.mysql.jdbc.Driver");
+                connection = DriverManager.getConnection(jdbcURL, jdbcUsername, jdbcPassword);
+            } catch (SQLException | ClassNotFoundException e) {
+                e.printStackTrace();
+            }
+        }
+        return connection;
+    }
+}
