+++ src/main/java/service/INewSpaperService.java
@@ -1,0 +1,17 @@
+package service;
+
+import model.NewSpaper;
+
+import java.util.List;
+
+public interface INewSpaperService {
+//    void add (NewSpaper newSpaper);
+//
+//    NewSpaper finByID(int id);
+//
+//    void update (NewSpaper newSpaper);
+//
+//    void delete(NewSpaper newSpaper);
+
+    List<NewSpaper> listNew();
+}
