+++ src/main/java/service/NewSpaperimpl/NewSpaperService.java
@@ -1,0 +1,39 @@
+package service.NewSpaperimpl;
+
+import model.NewSpaper;
+import repository.INewSpaperRepository;
+import repository.newSpaperimpl.NewSpaperRepository;
+import service.INewSpaperService;
+
+import java.util.List;
+
+public class NewSpaperService implements INewSpaperService {
+
+    INewSpaperRepository iCharityRepository = new NewSpaperRepository();
+
+//    @Override
+//    public void add(NewSpaper newSpaper) {
+//        iCharityRepository.add(newSpaper);
+//    }
+//
+//    @Override
+//    public NewSpaper finByID(int id) {
+//        return iCharityRepository.finByID(id);
+//    }
+//
+//
+//    @Override
+//    public void update(NewSpaper newSpaper) {
+//        iCharityRepository.update(newSpaper);
+//    }
+//
+//    @Override
+//    public void delete(NewSpaper newSpaper) {
+//        iCharityRepository.delete(newSpaper);
+//    }
+
+    @Override
+    public List<NewSpaper> listNew() {
+        return iCharityRepository.listNew();
+    }
+}
